+++ ReactiveSk-skript-v2_6_3/src/main/kotlin/com/github/tanokun/reactivesk/v263/skript/runtime/invoke/instantiation/mediator/RuntimeConstructorMediator.kt
@@ -1,4 +1,4 @@
-package com.github.tanokun.reactivesk.v263.skript.runtime.instantiation.mediator
+package com.github.tanokun.reactivesk.v263.skript.runtime.invoke.instantiation.mediator
 
 import org.bukkit.event.Event
 import org.bukkit.event.HandlerList
