+++ ReactiveSk-skript-v2_7_0/src/main/kotlin/com/github/tanokun/reactivesk/v270/ReactiveSkAddon.kt
@@ -6,12 +6,12 @@ import com.github.tanokun.reactivesk.v263.ClassesRegister
 import com.github.tanokun.reactivesk.v263.skript.analyze.ast.*
 import com.github.tanokun.reactivesk.v263.skript.resolve.variable.LocalTypedVariableDeclarationEffect
 import com.github.tanokun.reactivesk.v263.skript.runtime.array.TransformSingleTypeArrayExpression
-import com.github.tanokun.reactivesk.v263.skript.runtime.function.FunctionReturnEffect
-import com.github.tanokun.reactivesk.v263.skript.runtime.function.call.NonSuspendCallFunctionEffect
-import com.github.tanokun.reactivesk.v263.skript.runtime.function.call.NonSuspendCallFunctionExpression
-import com.github.tanokun.reactivesk.v263.skript.runtime.instantiation.InstantiationExpression
-import com.github.tanokun.reactivesk.v263.skript.runtime.instantiation.ResolveFieldValueEffect
+import com.github.tanokun.reactivesk.v263.skript.runtime.invoke.function.call.NonSuspendCallFunctionEffect
+import com.github.tanokun.reactivesk.v263.skript.runtime.invoke.function.call.NonSuspendCallFunctionExpression
+import com.github.tanokun.reactivesk.v263.skript.runtime.invoke.instantiation.InstantiationExpression
+import com.github.tanokun.reactivesk.v263.skript.runtime.invoke.instantiation.ResolveFieldValueEffect
 import com.github.tanokun.reactivesk.v263.skript.runtime.observe.ObserverSkriptEvent
+import com.github.tanokun.reactivesk.v263.skript.runtime.returns.FunctionReturnEffect
 import com.github.tanokun.reactivesk.v263.skript.runtime.variable.CastExpression
 import com.github.tanokun.reactivesk.v263.skript.runtime.variable.field.GetTypedValueFieldExpression
 import com.github.tanokun.reactivesk.v263.skript.runtime.variable.field.SetTypedValueFieldEffect
